Macro-generate `VMStackLimits`'s offsets/accessors/definition
This moves `VMStackLimits` into `for_each_vm_type!` so we can use
macro-generated offsets and alias region helpers instead of manually writing
them.
